+++ webapp/e2e-tests/electron/for_feature_development_not_LT_verification/graph_folder_loading/electron-unified-folder-loading/types.ts
@@ -1,0 +1,17 @@
+import type { ElectronApplication, Page } from '@playwright/test';
+import type { Core as CytoscapeCore } from 'cytoscape';
+import type { ElectronAPI } from '@/shell/electron';
+
+export interface ExtendedWindow {
+  cytoscapeInstance?: CytoscapeCore;
+  electronAPI?: ElectronAPI;
+}
+
+export type UnifiedFolderLoadingFixtures = {
+  electronApp: ElectronApplication;
+  appWindow: Page;
+  testProjectPath: string;
+  primaryVaultPath: string;
+  secondVaultPath: string;
+  tempUserDataPath: string;
+};
